+++ check_workflow/workflow.py
@@ -1,12 +1,13 @@
+import datetime as dt
 import string
 import typing as t
 from collections import defaultdict
 from pathlib import Path
 
-import yaml
 from gql.client import AsyncClientSession
 from packaging.specifiers import SpecifierSet
 from prettytable import PrettyTable, TableStyle
+from ruamel.yaml import YAML
 
 from check_workflow import WORKFLOW_T
 from check_workflow.gh_api import Release, fetch_releases
@@ -67,7 +68,8 @@ def extract_workflow_dependencies(raw_workflow: str) -> list[JobDependency]:
     NOTE: Only versioned actions are considered. Other specifications, such as local or docker
     actions, are skipped.
     """
-    loaded = yaml.safe_load(raw_workflow)
+    yaml_parser = YAML(typ="safe")
+    loaded = yaml_parser.load(raw_workflow)
 
     extracted_dependencies = []
     for job, job_params in loaded["jobs"].items():
@@ -99,7 +101,9 @@ class OutdatedDep(t.NamedTuple):  # noqa: D101
 
 
 async def report_outdated(
-    session: AsyncClientSession, raw_workflows: WORKFLOW_T
+    session: AsyncClientSession,
+    raw_workflows: WORKFLOW_T,
+    cooldown: dt.timedelta | None,
 ) -> dict[str, list[OutdatedDep]]:
     """Parse the provided workflow files and return a per-file list of outdated dependencies."""
     # Cache latest release info to cut down on API calls; keyed by (owner, repo) tuples
@@ -113,7 +117,10 @@ async def report_outdated(
             if cache_key not in seen_releases:
                 latest = (
                     await fetch_releases(
-                        session=session, owner=dep.uses.owner, repo_name=dep.uses.repo
+                        session=session,
+                        owner=dep.uses.owner,
+                        repo_name=dep.uses.repo,
+                        cooldown=cooldown,
                     )
                 )[0]
                 seen_releases[cache_key] = latest